+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - Per-AC/DC, per-profile tunings enabled (Battery vs AC power +  platform profile)
+- Add ability to restore PPT defaults (WIP)
 
 ## [v6.1.0-rc4]